Update is finished
Essentially added a variety of functions one that does random encounters
and one that randomizes weapons. Minor things that need to be done:
weapons list needs to be completed, encounters need to be customized,
more monsters need to be created.

Major things to be done: GUI needs to be integrated into the main file
of the game, combat needs to be reworked around the GUI, flags need to
be implemented into combat and travel to start creating the game, basic
story for each starting city must be made, quests in every town and
city, NPCS, lots to do!!!